refactor(studio): migrate admin router onto registry (phase 2 area 2)
Move all /admin/* handlers from routers/admin.py into services/admin.py
decorated with @studio_route. Service functions (doctor, health_report,
transition_sessions, etc.) are preserved verbatim; route handlers are
renamed with _route suffix to avoid collision, with name= overrides
preserving the original FastAPI route names.

Oracle: routes.json identical, openapi.json identical (79 routes).
