+++ lionagi/studio/app.py
@@ -14,7 +14,6 @@ from starlette.staticfiles import StaticFiles
 from .config import CORS_ORIGINS, HOST
 from .registry import iter_studio_routes, load_studio_route_modules
 from .routers import (
-    admin,
     schedules,
 )
 from .services import stats as stats_svc
@@ -153,7 +152,6 @@ for _route in iter_studio_routes():
         include_in_schema=_route.include_in_schema,
     )
 
-app.include_router(admin.router, prefix="/api")
 app.include_router(schedules.router, prefix="/api")
